Move solver module back to top level
