+++ scico/optimize/__init__.py
@@ -14,18 +14,8 @@ from ._admm import ADMM
 from ._ladmm import LinearizedADMM
 from ._pgm import PGM, AcceleratedPGM
 from ._primaldual import PDHG
-from ._solver import minimize, minimize_scalar, cg
 
-__all__ = [
-    "ADMM",
-    "LinearizedADMM",
-    "PGM",
-    "AcceleratedPGM",
-    "PDHG",
-    "minimize",
-    "minimize_scalar",
-    "cg",
-]
+__all__ = ["ADMM", "LinearizedADMM", "PGM", "AcceleratedPGM", "PDHG"]
 
 # Imported items in __all__ appear to originate in top-level linop module
 for name in __all__:
