+++ scico/optimize/admm.py
@@ -23,8 +23,8 @@ from scico.linop import CircularConvolve, Identity, LinearOperator
 from scico.loss import SquaredL2Loss, WeightedSquaredL2Loss
 from scico.math import is_real_dtype
 from scico.numpy.linalg import norm
-from scico.optimize._solver import cg as scico_cg
-from scico.optimize._solver import minimize
+from scico.solver import cg as scico_cg
+from scico.solver import minimize
 from scico.typing import JaxArray
 from scico.util import ensure_on_device
