+++ scico/solver.py
